web-animations/animation-sequence/animation-sequence-clone.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationSequence clone() method test</title>
<meta name="assert" content="Creates a deep copy of this AnimationSequence object">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationsequence-clone">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var source = new AnimationSequence([]);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationSequence object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationSequence object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationSequence.children ' +
        'should not be the same object as source.children');
    assert_array_equals(cloned.children, [], 'Cloned AnimationSequence.children should be empty');

    assert_false(cloned.timing === source.timing, 'Cloned AnimationSequence.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationSequence.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationSequence.clone() creates a copy of AnimationSequence object without children ' +
    'and with default timing');

test(function() {
    var source = new AnimationSequence([], 5);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationSequence object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationSequence object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationSequence.children ' +
        'should not be the same object as source.children');
    assert_array_equals(cloned.children, [], 'Cloned AnimationSequence.children should be empty');

    assert_false(cloned.timing === source.timing, 'Cloned AnimationSequence.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationSequence.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationSequence.clone() creates a copy of AnimationSequence object without children ' +
    'and with non-default timing duration');

test(function() {
    var timing = {
        delay: 1,
        endDelay: 1,
        fill: 'none',
        iterationStart: 1.1,
        iterations: 2.0,
        duration: 3,
        playbackRate: -2,
        playbackDirection: 'reverse',
        direction: 'reverse',
        easing: 'ease'};
    var source = new AnimationSequence([], timing);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationSequence object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationSequence object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationSequence.children ' +
        'should not be the same object as source.children');
    assert_array_equals(cloned.children, [], 'Cloned AnimationSequence.children should be empty');

    assert_false(cloned.timing === source.timing, 'Cloned AnimationSequence.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationSequence.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationSequence.clone() creates a copy of AnimationSequence object without children ' +
    'and with non-default timing');

test(function() {
    var node = newAnimation(createDiv(this));
    var children = [node];
    var source = new AnimationSequence(children);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationSequence object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationSequence object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationSequence.children ' +
        'should not be the same object as source.children');
    assert_equals(cloned.children.length, source.children.length, 'Cloned AnimationSequence ' +
        'object should contain the same number of children');
    assert_false(cloned.clidren[0] === source.children[0], 'First child of cloned ' +
        'AnimationSequence object should not be the same as first child of source');
    assert_object_equals(cloned.clidren[0], source.children[0], 'First child of cloned ' +
        'AnimationSequence object should be identical to first child of source');
}, 'Test AnimationSequence.clone() creates a copy of AnimationSequence object including children');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationSequence([node1,node2]);
    var node4 = newAnimation(createDiv(this));
    var node5 = newAnimation(createDiv(this));
    var node6 = new AnimationGroup([node4, node5]);

    var source = new AnimationSequence([node3, node6]);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationSequence object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationSequence object should be ' +
        'identical to source object');

    var expected = [node1, node2, node3, node4, node5, node6];
    var actual = [
        cloned.children[0].children[0],
        cloned.children[0].children[1],
        cloned.children[0],
        cloned.children[1].children[0],
        cloned.children[1].children[1],
        cloned.children[1]
    ];
    for (var i = 0; i < expected.length; i++) {
        assert_false(actual[i] === expected[i], 'Descendant of cloned AnimationSequence ' +
            'object should be a copy of corresponding descendant of source object');
        assert_object_equals(actual[i], expected[i], 'Descendant of cloned AnimationSequence ' +
            'object should be identical to corresponding descendant of source object');
    }
    assert_false(cloned.timing === source.timing, 'Cloned AnimationSequence.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationSequence.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationSequence.clone() creates a deep copy of AnimationSequence object');

test(function() {
    var node1 = new AnimationSequence([]);
    var node2 = new AnimationSequence([node1]);

    var cloned = node1.clone();
    assert_equals(cloned.parent, null, 'Cloned AnimationSequence.parent should be null');
    assert_equals(node1.parent, node2, 'Source AnimationSequence.parent should not be affected');
}, 'Test AnimationSequence.clone() set parent attribute to null on cloned object');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationSequence([]);
    var node3 = new AnimationSequence([node1, node2]);

    var cloned = node2.clone();
    assert_equals(cloned.previousSibling, null,
        'Cloned AnimationSequence.previousSibling should be null');
    assert_equals(node2.previousSibling, node1,
        'Source AnimationSequence.previousSibling should not be affected');
}, 'Test AnimationSequence.clone() set previousSibling attribute to null on cloned object');

test(function() {
    var node1 = new AnimationSequence([]);
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationSequence([node1, node2]);

    var cloned = node1.clone();
    assert_equals(cloned.nextSibling, null,
        'Cloned AnimationSequence.nextSibling should be null');
    assert_equals(node1.nextSibling, node2,
        'Source AnimationSequence.nextSibling should not be affected');
}, 'Test AnimationSequence.clone() set nextSibling attribute to null on cloned object');
</script>
</body>
